requirements.txt
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
flask==3.0.2
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-core==2.16.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
